Make admin added, Update profile added. CMS done.

+++ app/Http/Controllers/CategoriesController.php
@@ -96,6 +96,11 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Category can not be deleted because it has some post.');
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully.');
